# apidefinition.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0_60",
    "title": "IPS MERN API",
    "description": "OpenAPI 3.0 specification for the IPS MERN API."
  },
  "servers": [
    {
      "url": "https://ipsmern-dep.azurewebsites.net",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:5050",
      "description": "Local Server"
    }
  ],
...
  "paths": {
    "/ips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create a new IPS record",
        "description": "Create a new IPS record.",
        "requestBody": {
          "description": "MongoDb JSON object representing the new IPS record.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON object of the created IPS record."
          }
        }
      }
    },
    "/ipsmany": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create multiple IPS records",
        "description": "Create multiple IPS records.",
        "requestBody": {
          "description": "Array of MongoDb JSON objects representing the new IPS records.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of MongoDb JSON objects of the created IPS records."
          }
        }
      }
    },
    "/ipsbundle": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create IPS records from an IPS Bundle",
        "description": "Create IPS records from an IPS Bundle.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/pushipsora": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Push IPS data to an external ORA system",
        "description": "Push IPS data to an external ORA system.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response from the ORA system."
          }
        }
      }
    },
    "/pushipsnld": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Push IPS data to an external NLD system",
        "description": "Push IPS data to an external NLD system.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response from the NLD system."
          }
        }
      }
    },
    "/ipsfrombeer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create MongoDb IPS records from BEER format",
        "description": "Create MongoDb IPS records from BEER format.",
        "requestBody": {
          "description": "BEER - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/ipsfromcda": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create MongoDb IPS records from CDA XML format",
        "description": "Create MongoDb IPS records from CDA XML format.",
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/ipsfromhl72x": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create MongoDb IPS records from HL7 2.x format",
        "description": "Create MongoDb IPS records from HL7 2.x format.",
        "requestBody": {
          "description": "HL7 2.x - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/convertmongo2beer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert MongoDB format to BEER format",
        "description": "Convert MongoDB format to BEER format.",
        "requestBody": {
          "description": "MongoDB - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BEER - Plain Text"
          }
        }
      }
    },
    "/convertmongo2hl7": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert MongoDB format to HL7 2.3 format",
        "description": "Convert MongoDB format to HL7 2.3 format.",
        "requestBody": {
          "description": "MongoDB - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HL7 2.3 - Plain Text"
          }
        }
      }
    },
    "/convertbeer2mongo": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert BEER format to MongoDB format",
        "description": "Convert BEER format to MongoDB format.",
        "requestBody": {
          "description": "BEER - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDB JSON"
          }
        }
      }
    },
    "/convertbeer2ips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert BEER format to IPS JSON format",
        "description": "Convert BEER format to IPS JSON format.",
        "parameters": [
          {
            "$ref": "#/components/parameters/XIPSFormat"
          }
        ],
        "requestBody": {
          "description": "BEER - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPS Bundle JSON"
          }
        }
      }
    },
    "/convertips2beer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert IPS JSON format to BEER format",
        "description": "Convert IPS JSON format to BEER format.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BEER - Plain Text"
          }
        }
      }
    },
    "/convertips2plaintext": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert IPS JSON format to Plain Text Readable format",
        "description": "Convert IPS JSON format to Plain Text Readable format.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Human Readable - Plain Text"
          }
        }
      }
    },
    "/convertcdatoips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert CDA XML format to IPS JSON format",
        "description": "Convert CDA XML format to IPS JSON format.",
        "parameters": [
          {
            "$ref": "#/components/parameters/XIPSFormat"
          }
        ],
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPS Bundle JSON"
          }
        }
      }
    },
    "/convertcdatobeer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert CDA XML format to BEER format",
        "description": "Convert CDA XML format to BEER format.",
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BEER - Plain Text"
          }
        }
      }
    },
    "/convertcdatomongo": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert CDA XML format to MongoDB format",
        "description": "Convert CDA XML format to MongoDB format.",
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDB JSON"
          }
        }
      }
    },
    "/converthl72xtomongo": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert HL7 2.x format to MongoDB format",
        "description": "Convert HL7 2.x format to MongoDB format.",
        "requestBody": {
          "description": "HL7 2.x - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDB JSON"
          }
        }
      }
    },
    "/converthl72xtoips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert HL7 2.x format to IPS JSON format",
        "description": "Convert HL7 2.x format to IPS JSON format.",
        "parameters": [
          {
            "$ref": "#/components/parameters/XIPSFormat"
          }
        ],
        "requestBody": {
          "description": "HL7 2.x - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPS Bundle JSON"
          }
        }
      }
    },
    "/convertxml": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Generic convert XML format to JSON format",
        "description": "Generic convert XML format to JSON format.",
        "requestBody": {
          "description": "XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON"
          }
        }
      }
    },
    "/convertfhirxml": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert FHIR XML format to FHIR JSON format",
        "description": "Convert FHIR XML format to FHIR JSON format.",
        "requestBody": {
          "description": "FHIR XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "FHIR JSON"
          }
        }
      }
    },
    "/ipsUniVal": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Validates a IPS Unified format record.",
        "description": "Validates a IPS Unified format record.",
        "requestBody": {
          "description": "MongoDb JSON object representing the IPS Unified format record.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON object containing the validation report."
          }
        }
      }
    },
...
    "/ipsraw/{id}": {
      "get": {
        "tags": [
          "IPS"
        ],
        "summary": "Retrieve raw IPS record by ID",
        "description": "Retrieve raw IPS record by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "IPS record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object of the raw IPS record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/ipsmongo/{id}": {
      "get": {
        "tags": [
          "IPS"
        ],
        "summary": "Retrieve IPS record in MongoDB format by ID",
        "description": "Retrieve IPS record in MongoDB format by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MongoDB JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/ips/{id}": {
      "get": {
        "tags": [
          "IPS"
        ],
        "summary": "Retrieve IPS record in expanded format by ID",
        "description": "Retrieve IPS record in expanded format by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expanded IPS - FHIR JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "IPS"
        ],
        "summary": "Update an IPS record by ID",
        "description": "Update an IPS record by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "IPS record ID"
          }
        ],
        "requestBody": {
          "description": "Partial JSON object representing the updated IPS record.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated IPS record in MongoDB JSON format.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "IPS"
        ],
        "summary": "Delete an IPS record by ID",
        "description": "Delete an IPS record by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmation of deletion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ipsbasic/{id}": {
      "get": {
        "tags": [
          "IPS"
        ],
        "summary": "Retrieve IPS record in basic format by ID",
        "description": "Retrieve IPS record in basic format by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Basic format - Plain Text",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ipsplaintext/{id}": {
      "get": {
        "tags": [
          "IPS"
        ],
        "summary": "Retrieve IPS record in human readable format by ID",
        "description": "Retrieve IPS record in human readable format by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Human Readable format - Plain Text",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ipsbeer/{id}/{delim}": {
      "get": {
        "tags": [
          "IPS"
        ],
        "summary": "Retrieve IPS record in BEER format by ID with optional delimiter",
        "description": "Retrieve IPS record in BEER format by ID with optional delimiter. The delimiter options are documented in the API docs.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "delim",
            "in": "path",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional delimiter. Available options: 'semi' (;), 'colon' (:), 'pipe' (|), 'at' (@), 'newline' (\\n)."
          }
        ],
        "responses": {
          "200": {
            "description": "BEER - Plain Text",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "XIPSFormat": {
        "name": "x-ips-format",
        "in": "header",
        "description": "Optional format override. Available values: 'unified' (default), 'inter', 'legacy'.",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
  }
}